Fix program codegen for conversions between primitive types
While trying to update pulumi-dotnet in pulumi/pulumi, a smoke test failed https://github.com/pulumi/pulumi/pull/24249

```
Error:         D:\a\_temp\test-env3863622462\out\Rc.cs(29,25): error CS0029: Cannot implicitly convert type 'Pulumi.Output<string>' to 'Pulumi.Output<dynamic>' [D:\a\_temp\test-env3863622462\out\test-env3863622462.csproj]
```

We need to properly handle the IDType in codegen that was introduced in 3.256.0.

Getting `l2-id-type` passing required dealing with primitive conversions, so that solved a couple other tests along with it.

Int to number conversions now always emit a `(double)` cast because C#'s implicitly-typed arrays infer int[] from bare int elements, but  that doesn't convert to InputList<double>. Codegen can't tell scalar from array-element context, so we just always cast.

+++ pulumi-language-dotnet/codegen/testdata/third-party-package-pp/third-party-package.cs
@@ -12,12 +12,12 @@ return await Deployment.RunAsync(() =>
 
     var Question = new Other.Module.Object("Question", new()
     {
-        Answer = 42,
+        Answer = (double)42,
     });
 
     var Question2 = new Other.Module.Sub.Object("Question2", new()
     {
-        Answer = 24,
+        Answer = (double)24,
     });
 
     var Provider = new Other.Provider("Provider", new()
